feat: 공모전 상세 내용 제작 완료
