+++ teamkerbell_frontend/src/components/mainComponents/CompDetail.module.css
@@ -1,8 +1,21 @@
-.compDetail {
+@font-face {
+    font-family: "NanumGothic"; /* regular */
+    font-weight: normal;
+    src: url("../../../public/fonts/NanumGothic.ttf") format("truetype");
+}
 
+@font-face {
+    font-family: "NanumSquareNeo-cBd"; /* Bold */
+    font-weight: normal;
+    src: url("../../../public/fonts/NanumSquareNeo-cBd.ttf") format("truetype");
+}
+
+
+
+.compDetail {
     padding-top: 100px;
-    padding-left: 100px;
-    padding-right: 20px;
+    padding-left: 200px;
+    padding-right: 200px;
     margin: 10px;
     width: 90%;
     text-align: center; 
@@ -11,26 +24,49 @@
 
   
 .image {
-  float:left;
+  margin-top: 40px;
   padding-left: 100px;
-  width: 500px;
-  height: 500px;
+  max-width: 100%;
+  height:auto;
   vertical-align: middle;
 }
 
 .title {
   text-align: center;
   vertical-align: middle;
-  font-size: 2rem;
-  margin-top: 20px;
+  font-size: 3rem;
+  font-family: "NanumSquareNeo-cBd";
+  margin-top: 15px;
   color: #044018;
 }
 
 .description {
-  float:right;
+  vertical-align: middle;
   margin-top: 10px;
-  font-size:1rem;
+  font-size:1.8rem;
+  font-family:"NanumGothic";
+  line-height: 2;
 }
 
+.daycount{
+  font-weight: bold;
+  font-style: italic;
+  color: #ff0000;
+  display:inline;
+  margin-top: 20px;
+  margin-left: 50px;
+  padding: 2px 6px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  background-color: var(--Fill-Strong, #70737c29);
+}
+
+
+.grid {
+    display:grid;
+    gap:200px;
+    grid-template-columns:  minmax(400px, 1fr) 1.5fr;
+    justify-content: center;
+}
